requirements.txt
streamlit
pandas
numpy
plotly
google-cloud-bigquery
google-auth
db-dtypes
pandas-gbq